Move calculate tanimoto score functions to calculate tanimoto scores
